# cart-service/openapi.yaml
openapi: 3.0.3
info:
  title: cart-service
  version: 1.0.0

servers:
  - url: http://localhost:8083

paths:
  /api/cart:
    get:
      summary: Get user's cart
      description: >
        Supply the user via X-User-Id header (from API Gateway) and/or userId query parameter.
        At least one is required. If both are sent, they must match.
      operationId: getCart
      parameters:
        - name: X-User-Id
          in: header
          required: false
          schema:
            type: string
          description: Authenticated user id (forwarded by API Gateway)
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: Same user id as an alternative to the header (must match X-User-Id when both are used)
      responses:
        '400':
          description: Missing user identification or header/query mismatch
        '200':
          description: Cart details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
    delete:
      summary: Clear cart
      operationId: clearCart
      parameters:
        - $ref: '#/components/parameters/XUserId'
      responses:
        '204':
          description: Cart cleared

  /api/cart/items:
    post:
      summary: Add item to cart
      operationId: addItem
      parameters:
        - $ref: '#/components/parameters/XUserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemInput'
      responses:
        '201':
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'

  /api/cart/items/{productId}:
    put:
      summary: Update item quantity
      operationId: updateItem
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XUserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuantityRequest'
      responses:
        '200':
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
    delete:
      summary: Remove item from cart
      operationId: removeItem
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XUserId'
      responses:
        '200':
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'

components:
  parameters:
    XUserId:
      name: X-User-Id
      in: header
      required: true
      schema:
        type: string

  schemas:
    CartItemInput:
      type: object
      required:
        - productId
        - productName
        - price
        - quantity
      properties:
        productId:
          type: string
        productName:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer

    UpdateQuantityRequest:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer

    CartItem:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer

    CartResponse:
      type: object
      properties:
        cartId:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        totalAmount:
          type: number
          format: double
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time